refactor to move everything within the page
