--- wtm-react/src/pages/eventcreatorpage.jsx
@@ -1,9 +1,0 @@
-import EventCreator from '../components/EventCreation/eventcreator'; // No curly braces for default export
-  
-function EventCreatorPage() {
-  return (
-    <EventCreator />
-  );
-}
-  
-export default EventCreatorPage;
