+++ wtm-react/src/pages/eventdetails.jsx
@@ -1,0 +1,13 @@
+
+/*
+  TODO
+  - Make a page for event creator, or event participant
+*/ 
+
+function EventDetails() {
+  return (
+    <div>EventDetails</div>
+  )
+}
+  
+export default EventDetails;
