--- wtm-react/src/pages/eventdetailspage.jsx
@@ -1,14 +1,0 @@
-import EventDetails from "../components/EventDetails/eventdetails";
-
-/*
-  TODO
-  - Make a page for event creator, or event participant
-*/ 
-
-function EventDetailsPage() {
-    return (
-      <EventDetails />
-    );
-  }
-  
-  export default EventDetailsPage;
