+++ wtm-react/src/pages/homeview.jsx
@@ -1,0 +1,15 @@
+import JoinedEvents from "../components/HomeView/joinedevents";
+import HomepageHeader from "../components/HomeView/homepageheader";
+import FriendsDisplay from "../components/HomeView/friendsdisplay";
+
+function HomeView() {
+    return (
+      <div className='HomeView'>
+        <HomepageHeader />
+        <JoinedEvents />
+        <FriendsDisplay />
+      </div>
+    );
+  }
+  
+  export default HomeView;
