--- wtm-react/src/pages/homeviewpage.jsx
@@ -1,9 +1,0 @@
-import HomeView from '../components/HomeView/homeview';
-
-function HomeViewPage() {
-    return (
-      <HomeView />
-    );
-  }
-  
-  export default HomeViewPage;
